Add deck service
