+++ dartmuti/lib/tabletop_component.dart
@@ -2,15 +2,18 @@ import 'package:angular2/angular2.dart';
 import 'tabletop.dart';
 import 'trick_component.dart';
 import 'card_component.dart';
+import 'deck_service.dart';
 
 @Component(
     selector: 'dartmuti-tabletop',
     templateUrl: 'tabletop_component.html',
+    providers: const [DeckService],
     directives: const [NgFor, NgIf, TrickComponent, CardComponent])
 class TabletopComponent {
   Tabletop model;
-  TabletopComponent() {
+  TabletopComponent(DeckService DS) {
     model = new Tabletop('Test state');
+    model.DeckService = DS;
     model.init();
   }
 }
